Hourly weather data for 50 US cities.

A dataset containing hourly weather data for 50 US cities.

Usage

weather

Format

A data frame with 6000 rows and 21 variables:

city

Name of city.

state

Name of state.

time

Time.

summary

A human-readable text summary.

icon

A machine-readable text summary, suitable for selecting an icon for display.

precip_intensity

The intensity (in inches of liquid water per hour) of precipitation occurring at the given time.

precip_probability

The probability of precipitation occurring, between 0 and 1, inclusive.

temperature

The air temperature in degrees Fahrenheit.

apparent_temperature

The apparent (or “feels like”) temperature in degrees Fahrenheit.

dew_point

The dew point in degrees Fahrenheit.

humidity

The relative humidity, between 0 and 1, inclusive.

pressure

The sea-level air pressure in millibars.

wind_speed

The wind speed in miles per hour.

wind_gust

The time at which the maximum wind gust speed occurs during the day.

wind_bearing

The direction that the wind is coming from in degrees, with true north at 0° and progressing clockwise. (If windSpeed is zero, then this value will not be defined.)

cloud_cover

The percentage of sky occluded by clouds, between 0 and 1, inclusive.

uv_index

The UV index.

visibility

The average visibility in miles, capped at 10 miles.

ozone

The columnar density of total atmospheric ozone at the given time in Dobson units.

precip_type

The type of precipitation occurring at the given time. If defined, this property will have one of the following values: "rain", "snow", or "sleet".

forecast

Indicator for whether measurement is historical (FALSE) or forecast (TRUE).

Source

Dark Sky, https://darksky.net/dev